+++ lib/src/positioned/positioned.dart
@@ -1,5 +1,3 @@
-export 'animation/positioned_animation.dart';
-export 'dismissible/positioned_dismiss.dart';
 export 'shell/positioned_shell.dart';
 export 'dialog/positioned_dialog.dart'
     hide PositionedDialogInsert, PositionedDialogRemove;